pycryptodomex/pyproject.toml
[project]

name = "pycryptodomex_app"
version = "1.0.0"
description = ""
authors = [{name = "Flet Developer", email = "your@email.com"}]
dependencies = ["flet==0.25.1", "pycryptodomex"]

[tool.flet]
android.adaptive_icon_background = "#ffffff"